Indulging in some refactoring.
Changing the default CRS to degrees at 1° in EPSG:4326.

Some cleanup of Xee internals. Mainly, we ensure that internally, "width"s are associated with the "x" dim, and "heights" with "y".

This change includes a few minor fixes: it adds a few more sanity checks for valid bounds. More thorough testing for parity between pure `computePixels()` and Xee is needed. Further, we fix a bug in the micro benchmarks (wrong dimension name).

PiperOrigin-RevId: 566319203