Xee support for distributed worker initialization/authentication
This adds an Earth Engine initialization check for dataset operations for remote workers to call Earth Engine. Also adds docs for submitting a Dataflow job using Xee

Dataflow jobs would fail with Xee due to the remote workers not having the EE client library initialized. This adds a check to all calls for the `EarthEngineBackendArray` object so that if there is a call to EE, it will be initialized if not already. It also includes a parameter a user has to set to allow the attempt of automatically initializing Earth Engine for workers, this is so users are explicit and know that EE will try initializing.

There was discussion on issue #99 regarding documentation for how to do initialize/authenticate on distributed cluster and this also includes a Dataflow example (docs, cloud setup, script) where that users can start from.

close #51

PiperOrigin-RevId: 604388395