set the version number somewhere
Fix #161 

My understanding of the problem is that the version number is not set anywhere. 

pyproject.toml uses a dynamic version number but the dynamic estination is not set, xee package has no __version__ variable. the one in `ext` is using the one from metadata which relies on the pyproject file. 

I decided to start this PR by setting where the dynamic version number should come from. Now I need your review to know where  the hard coded version number should be set.