Fix references to deprecated Landsat Collection 1.
This fixes broken tests in xee.ext_integration_test.py

The tests were broken because the Landsat collection LANDSAT/LC08/C01/T1 was removed from Earth Engine. PR #157 was supposed to fix this but missed some references to "C01". This replaces all references to "C01" with "C02" which is available in the data catalog.

PiperOrigin-RevId: 656004125