use the python default gitignore
I simply used the python global gitignore and removed everything related to the wheel builder that are not related to this project (namely poetry, hatch, pdm) and things related to pycharm. 

I added back the MacOs `.DS_Store` at the bottom. 

I did this because in another PR I added stuff related to nox and all folder went in the last commit. I think it cost nothing to have a wider gitignore and make sure that new tools can be brought to the package without ignore issues.